Replace httpserver.c

+++ asgn4/httpserver.c
@@ -124,9 +124,9 @@ int main(int argc, char **argv) {
 void *worker_threads(){
     while(true){
         fprintf(stderr, "thread  is in\n");
-        int conn = 0;
+        uintptr_t conn = 0;
         queue_pop(task_queue,(void **)&conn);
-        fprintf(stderr, "conn: %d\n", *(int*)&conn);
+        fprintf(stderr, "conn: %lu\n", *(uintptr_t*)&conn);
         fprintf(stderr, "queue pop is smooth\n");
         //handle_connection(connfd);
         //close(connfd);
